refactor: split CancellationTokenKustoExtensions into focused classes
