+++ src/Atc.Kusto/Extensions/Internal/CslAdminProviderExtensions.cs
@@ -2,10 +2,9 @@
 namespace Atc.Kusto.Extensions.Internal;
 
 /// <summary>
-/// Utilities to hook up server-side cancellation for Kusto operations by issuing a cancel control command
-/// when a CancellationToken is triggered.
+/// Extension methods for <see cref="ICslAdminProvider"/> to support server-side cancellation.
 /// </summary>
-internal static class CancellationTokenKustoExtensions
+internal static class CslAdminProviderExtensions
 {
     /// <summary>
     /// Determines whether server-side cancellation should be enabled based on the admin provider and options.
@@ -82,36 +81,6 @@ internal static class CancellationTokenKustoExtensions
                 }
             },
             state);
-    }
-
-    /// <summary>
-    /// Determines if an exception is a cancellation exception (including Kusto-specific ones).
-    /// </summary>
-    /// <param name="exception">The exception to check.</param>
-    /// <returns>True if the exception represents a cancellation; otherwise, false.</returns>
-    public static bool IsCancellationException(Exception exception)
-        => exception is OperationCanceledException
-           || exception is TaskCanceledException;
-
-    /// <summary>
-    /// Normalizes cancellation exceptions to standard OperationCanceledException.
-    /// </summary>
-    /// <param name="exception">The exception to normalize.</param>
-    /// <param name="cancellationToken">The cancellation token associated with the operation.</param>
-    /// <returns>A normalized OperationCanceledException.</returns>
-    public static OperationCanceledException NormalizeCancellationException(
-        Exception exception,
-        CancellationToken cancellationToken)
-    {
-        if (exception is OperationCanceledException oce)
-        {
-            return oce;
-        }
-
-        return new OperationCanceledException(
-            $"The operation was canceled. Original exception: {exception.GetType().Name}",
-            exception,
-            cancellationToken);
     }
 
     private sealed class NullDisposable : IDisposable
